1차

+++ app/main/service/privacy_service.py
@@ -1,6 +1,7 @@
 from main import db
 from main.model.user import User
 from main.model.user_subject import UserSubject
+from main.model.user_complete import UserComplete
 import string, random
 
 from ..service.mailer_service import sendmail
@@ -127,10 +128,18 @@ def dropout(user_email,data):
             db.session.close()
             
             favorite = UserSubject.query.filter_by(email=user_email).all()
-            for i in favorite:
-                db.session.delete(i)
-                db.session.commit()
-                db.session.close()
+            if favorite:
+                for i in favorite:
+                    db.session.delete(i)
+                    db.session.commit()
+                    db.session.close()
+            
+            compelete = UserComplete.query.filter_by(email=user_email).all()
+            if compelete:
+                for i in compelete:
+                    db.session.delete(i)
+                    db.session.commit()
+                    db.session.close()
 
             response_object = {
                 'status': 'success',
